fix(tests): cast recovery sample points to float64 for py3.10 mypy
Bare np.linspace is typed floating[Any] under py3.10's older numpy stubs,
which fails the strictly-typed quad() arg in the 1D restriction-recovery
test. Cast to float64, matching the existing pattern in
test_thb_validation_basis.py.

Refs #164.

+++ tests/test_thb_validation_identities.py
@@ -162,7 +162,7 @@ class TestCoarseningProjection:
 
         c_fine = l2_project_thb(fine, quad).coeffs
         c_back = fine.restriction_to(coarse) @ c_fine
-        xs = np.linspace(0.02, 0.98, 80).reshape(-1, 1)
+        xs = np.linspace(0.02, 0.98, 80).reshape(-1, 1).astype(np.float64)
         np.testing.assert_allclose(
             THBSpline(coarse, c_back).evaluate(xs).ravel(), quad(xs), atol=1e-8
         )
